+++ Classes/Stop.h
@@ -14,10 +14,10 @@
 class Stop {
 private:
     std::string name, code, zone;
-    long latitude, longitude;
+    double latitude, longitude;
 public:
     Stop();
-    Stop(std::string name, std::string code, std::string zone, long latitude, long longitude);
+    Stop(std::string name, std::string code, std::string zone, double latitude, double longitude);
 
     std::vector<std::vector<std::string>> read(std::string filepath) const;
     std::vector<std::string> read_stop(std::string line) const;
